// src/app/shared/models/addressrequest.ts
import { IPaging } from "./paging";

export interface IAddressRequest {
    AddressID: number;
    CustomerId: number;
    Type: string;
    Line1: string;
    Line2: string;
    Line3: string;
    City: string;
    State: string;
    Country: string;
    Zip1: string;
    Zip2: string;
    IsPreferred: boolean;
    UserName: string;
    IsActive: boolean;
    SubSystem: string;
    ActivitySource: string;
    UserId: number;
    LoginId: number;
    Paging: IPaging;
    IsActivityRequired: boolean;
    IsInvalidAddress: boolean;
    CheckBlockList:boolean;
}